backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
python-dotenv==1.0.1
sqladmin==0.16.1
httptools==0.6.1
websockets==12.0
greenlet==3.0.3